Refactor into lib folder
This application now follows the Ruby standard where everything is organized into the "lib" folder
