+++ lib/editor.rb
