+++ lib/hash.rb
