+++ lib/printer.rb
