+++ lib/string_helper.rb
