+++ lib/templates/favicon.svg.erb
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+<!DOCTYPE svg PUBLIC "-//W3C//DTD SVG 1.1//EN" "http://www.w3.org/Graphics/SVG/1.1/DTD/svg11.dtd">
+<svg width="128" height="128" viewBox="0 0 128 128" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
+    <rect fill="#fff" stroke="#000" x="0" y="0" width="128" height="128" stroke-width="4" />
+    <rect fill="#fff" stroke="#000" x="32" y="32" width="64" height="64" stroke-width="2" />
+    <text x="10" y="30" fill="green" font-size="32"><%= Date.today %></text>
+</svg>
