+++ lib/templates/index.html.erb
