+++ lib/worklogger.rb
